Fix Flake8
Looks like `flake8` was targetting the wrong location, causing
it to be a no-op.  This change fixes the configuration and a few
linter errors that creeped in while the linter was a no-op.

One of these changes involves moving module-level imports to the
top of the file, but this revealed circular references, so the
change also moves some things from the main `__init__.py` out
into other modules.
