+++ tox.ini
@@ -9,7 +9,7 @@ envlist =
 deps =
   -rtests/requirements.txt
 commands =
-  flake8 aiotk/ tests/
+  flake8 src/aiotk/ tests/
   coverage erase
   coverage run -m pytest {posargs:-s -vv tests/}
   coverage html
